feat: Complete gap analysis implementation — Recording, Scrubbing, DSP fixes, Providers
## Summary
Major feature completion addressing FluxForge gap analysis vs industry DAWs:

### Recording System (TIER 0 - BLOCKING)
- Integrated RecordingManager with PlaybackEngine
- Added Recording state to PlaybackState enum
- Lock-free ring buffer recording pipeline

### Scrubbing Audio Preview (TIER 0 - BLOCKING)
- Full scrub state machine with velocity tracking
- 50ms default scrub window for audio preview
- FFI exports: start_scrub, update_scrub, stop_scrub
- Dart/Flutter integration in timeline_playback_provider

### DSP Performance Fixes (TIER 1 - CRITICAL)
- reverb.rs: Replaced naive DFT O(n²) with rustfft O(n log n)
- Proper Complex<f64> FFT buffer handling
- biquad/dynamics confirmed correct (IIR requires serial processing)

### New Providers (Flutter)
- CompingProvider: Take system, comp editing, ranking
- VideoProvider: Video playback, frame-accurate sync, timecode
- MidiProvider: MIDI clips, notes, piano roll integration
- MidiClipWidget: Timeline MIDI visualization

### Plugin System
- VST3 platform-specific editor hosting (macOS/Windows/Linux)
- Documented GUI embedding requirements
- preferred_editor_size() methods

### Research & Analysis
- Gap analysis document vs Cubase/Pro Tools/Logic/REAPER
- Pro Tools 2024 complete analysis
- REAPER 7 features research

95 files changed, 7277 insertions(+), 1409 deletions(-)

+++ crates/rf-audio/src/dsd_output.rs
@@ -534,7 +534,7 @@ mod tests {
         let dop_samples: Vec<i32> = (0..16)
             .map(|i| {
                 let marker = if i % 2 == 0 { 0x05 } else { 0xFA };
-                (marker as i32) << 24
+                marker << 24
             })
             .collect();
